Enable Snapshots and change domain
https://github.com/creek-service/creek-test/issues/30
https://github.com/creek-service/creek-test/issues/24
