+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,10 +1,3 @@
-### Description
-_Link to the issue you are fixing_.
-_What behavior do you want to change, why, how does your patch achieve the changes?_
-
-### Testing done
-_Describe the testing strategy. Unit and integration tests are expected for any behaviour changes._
-
 ### Reviewer checklist
 - [ ] Ensure relevant issues are linked (description should include text like "Fixes #<issue number>")
 - [ ] Ensure any appropriate documentation has been added or amended
