--- src/main/java/module-info.java
@@ -1,3 +1,0 @@
-module creek.json.schema.gradle.plugin {
-    exports org.creek.api.json.schema.gradle.plugin;
-}
