--- src/test/java/module-info.test
@@ -1,14 +1,0 @@
---add-modules
-  org.junitpioneer,org.hamcrest,guava.testlib,creek.test.util,creek.test.hamcrest,creek.test.conformity
-
---add-reads
-  creek.json.schema.gradle.plugin=org.junitpioneer,org.hamcrest,guava.testlib,creek.test.util,creek.test.hamcrest,creek.test.conformity
-
---add-opens
-  org.junitpioneer/org.junitpioneer.jupiter=org.junit.platform.commons
-
---add-opens
-  java.base/java.lang=org.junitpioneer
-
---add-opens
-  java.base/java.util=org.junitpioneer
